+++ packages/core/src/hl7.ts
@@ -258,3 +258,38 @@ export class Hl7Field {
     );
   }
 }
+
+interface Hl7DateParseOptions {
+  seconds?: boolean;
+  tzOffset?: string;
+}
+
+/**
+ * Returns a formatted string representing the date in ISO-8601 format.
+ * @param hl7Date Date string.
+ * @param options Optional configuration Object
+ * @returns
+ */
+export function parseHl7Date(hl7Date: string | undefined, options?: Hl7DateParseOptions): string | undefined {
+  if (!hl7Date) {
+    return undefined;
+  }
+
+  options = { ...{ seconds: true, tzOffset: 'Z' }, ...options };
+
+  const year = Number.parseInt(hl7Date.substring(0, 4));
+  const month = Number.parseInt(hl7Date.substring(4, 6));
+  const date = Number.parseInt(hl7Date.substring(6, 8));
+  const hours = Number.parseInt(hl7Date.substring(8, 10));
+  const minutes = Number.parseInt(hl7Date.substring(10, 12));
+
+  const seconds = options.seconds ? Number.parseInt(hl7Date.substring(12, 14)) : 0;
+
+  return `${pad2(year)}-${pad2(month)}-${pad2(date)}T${pad2(hours)}:${pad2(minutes)}:${pad2(seconds)}.000${
+    options.tzOffset
+  }`;
+}
+
+function pad2(n: number): string {
+  return n.toString().padStart(2, '0');
+}
